+++ adbai-20250812/alibabacloud_adbai20250812/models/_create_embodied_aiplatform_request.py
@@ -14,6 +14,7 @@ class CreateEmbodiedAIPlatformRequest(DaraModel):
         device_count: int = None,
         platform_name: str = None,
         ray_config: main_models.CreateEmbodiedAIPlatformRequestRayConfig = None,
+        ray_train_config: main_models.CreateEmbodiedAIPlatformRequestRayTrainConfig = None,
         region_id: str = None,
         webserver_spec_name: str = None,
     ):
@@ -23,6 +24,7 @@ class CreateEmbodiedAIPlatformRequest(DaraModel):
         # This parameter is required.
         self.platform_name = platform_name
         self.ray_config = ray_config
+        self.ray_train_config = ray_train_config
         # This parameter is required.
         self.region_id = region_id
         self.webserver_spec_name = webserver_spec_name
@@ -30,6 +32,8 @@ class CreateEmbodiedAIPlatformRequest(DaraModel):
     def validate(self):
         if self.ray_config:
             self.ray_config.validate()
+        if self.ray_train_config:
+            self.ray_train_config.validate()
 
     def to_map(self):
         result = dict()
@@ -47,6 +51,9 @@ class CreateEmbodiedAIPlatformRequest(DaraModel):
 
         if self.ray_config is not None:
             result['RayConfig'] = self.ray_config.to_map()
+
+        if self.ray_train_config is not None:
+            result['RayTrainConfig'] = self.ray_train_config.to_map()
 
         if self.region_id is not None:
             result['RegionId'] = self.region_id
@@ -71,11 +78,184 @@ class CreateEmbodiedAIPlatformRequest(DaraModel):
             temp_model = main_models.CreateEmbodiedAIPlatformRequestRayConfig()
             self.ray_config = temp_model.from_map(m.get('RayConfig'))
 
+        if m.get('RayTrainConfig') is not None:
+            temp_model = main_models.CreateEmbodiedAIPlatformRequestRayTrainConfig()
+            self.ray_train_config = temp_model.from_map(m.get('RayTrainConfig'))
+
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
 
         if m.get('WebserverSpecName') is not None:
             self.webserver_spec_name = m.get('WebserverSpecName')
+
+        return self
+
+class CreateEmbodiedAIPlatformRequestRayTrainConfig(DaraModel):
+    def __init__(
+        self,
+        cpu_acu: int = None,
+        gpu_specs: List[main_models.CreateEmbodiedAIPlatformRequestRayTrainConfigGpuSpecs] = None,
+        terminal_config: main_models.CreateEmbodiedAIPlatformRequestRayTrainConfigTerminalConfig = None,
+    ):
+        self.cpu_acu = cpu_acu
+        self.gpu_specs = gpu_specs
+        self.terminal_config = terminal_config
+
+    def validate(self):
+        if self.gpu_specs:
+            for v1 in self.gpu_specs:
+                 if v1:
+                    v1.validate()
+        if self.terminal_config:
+            self.terminal_config.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.cpu_acu is not None:
+            result['CpuAcu'] = self.cpu_acu
+
+        result['GpuSpecs'] = []
+        if self.gpu_specs is not None:
+            for k1 in self.gpu_specs:
+                result['GpuSpecs'].append(k1.to_map() if k1 else None)
+
+        if self.terminal_config is not None:
+            result['TerminalConfig'] = self.terminal_config.to_map()
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('CpuAcu') is not None:
+            self.cpu_acu = m.get('CpuAcu')
+
+        self.gpu_specs = []
+        if m.get('GpuSpecs') is not None:
+            for k1 in m.get('GpuSpecs'):
+                temp_model = main_models.CreateEmbodiedAIPlatformRequestRayTrainConfigGpuSpecs()
+                self.gpu_specs.append(temp_model.from_map(k1))
+
+        if m.get('TerminalConfig') is not None:
+            temp_model = main_models.CreateEmbodiedAIPlatformRequestRayTrainConfigTerminalConfig()
+            self.terminal_config = temp_model.from_map(m.get('TerminalConfig'))
+
+        return self
+
+class CreateEmbodiedAIPlatformRequestRayTrainConfigTerminalConfig(DaraModel):
+    def __init__(
+        self,
+        acr_config: main_models.CreateEmbodiedAIPlatformRequestRayTrainConfigTerminalConfigAcrConfig = None,
+    ):
+        self.acr_config = acr_config
+
+    def validate(self):
+        if self.acr_config:
+            self.acr_config.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.acr_config is not None:
+            result['AcrConfig'] = self.acr_config.to_map()
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AcrConfig') is not None:
+            temp_model = main_models.CreateEmbodiedAIPlatformRequestRayTrainConfigTerminalConfigAcrConfig()
+            self.acr_config = temp_model.from_map(m.get('AcrConfig'))
+
+        return self
+
+class CreateEmbodiedAIPlatformRequestRayTrainConfigTerminalConfigAcrConfig(DaraModel):
+    def __init__(
+        self,
+        instance_id: str = None,
+        namespaces: List[str] = None,
+        registry: str = None,
+    ):
+        self.instance_id = instance_id
+        self.namespaces = namespaces
+        self.registry = registry
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.instance_id is not None:
+            result['InstanceId'] = self.instance_id
+
+        if self.namespaces is not None:
+            result['Namespaces'] = self.namespaces
+
+        if self.registry is not None:
+            result['Registry'] = self.registry
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('InstanceId') is not None:
+            self.instance_id = m.get('InstanceId')
+
+        if m.get('Namespaces') is not None:
+            self.namespaces = m.get('Namespaces')
+
+        if m.get('Registry') is not None:
+            self.registry = m.get('Registry')
+
+        return self
+
+class CreateEmbodiedAIPlatformRequestRayTrainConfigGpuSpecs(DaraModel):
+    def __init__(
+        self,
+        allocate_unit: str = None,
+        count: int = None,
+        spec_name: str = None,
+    ):
+        self.allocate_unit = allocate_unit
+        self.count = count
+        self.spec_name = spec_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.allocate_unit is not None:
+            result['AllocateUnit'] = self.allocate_unit
+
+        if self.count is not None:
+            result['Count'] = self.count
+
+        if self.spec_name is not None:
+            result['SpecName'] = self.spec_name
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AllocateUnit') is not None:
+            self.allocate_unit = m.get('AllocateUnit')
+
+        if m.get('Count') is not None:
+            self.count = m.get('Count')
+
+        if m.get('SpecName') is not None:
+            self.spec_name = m.get('SpecName')
 
         return self